Update cp -pr to cp -r in lib/testers/starter-code-tester.ts
This change updates the cp command in the doTest() function of lib/testers/starter-code-tester.ts file. It replaces cp -pr with cp -r to
copy the starter directory to the copied starter directory. The -p option is removed, which preserves the file attributes, such as
timestamps and permissions. The commit message for this change could be "Update cp command in starter-code-tester.ts" or "Fix cp command
in starter-code-tester.ts".